Reintegrate tests into modular structure
